requirements.txt
allure-pytest==2.13.5
pytest==8.3.2
selenium==4.23.0
requests==2.32.3
webdriver-manager==4.0.2
python-dotenv==1.0.1